feat: move listener to render module
