+++ packages/sheets-ui/src/controllers/render-controllers/format-painter.render-controller.ts
@@ -15,19 +15,24 @@
  */
 
 import type { Workbook } from '@univerjs/core';
+import type { IRenderContext, IRenderModule } from '@univerjs/engine-render';
 import {
     Disposable,
+    ICommandService,
     toDisposable,
 } from '@univerjs/core';
-import type { IRenderContext, IRenderModule } from '@univerjs/engine-render';
 import { CURSOR_TYPE } from '@univerjs/engine-render';
 
+import { ApplyFormatPainterCommand, SetOnceFormatPainterCommand } from '../../commands/commands/set-format-painter.command';
 import { FormatPainterStatus, IFormatPainterService } from '../../services/format-painter/format-painter.service';
+import { ISheetSelectionRenderService } from '../../services/selection/base-selection-render.service';
 
 export class FormatPainterRenderController extends Disposable implements IRenderModule {
     constructor(
         private readonly _context: IRenderContext<Workbook>,
-        @IFormatPainterService private readonly _formatPainterService: IFormatPainterService
+        @IFormatPainterService private readonly _formatPainterService: IFormatPainterService,
+        @ISheetSelectionRenderService private readonly _selectionRenderService: ISheetSelectionRenderService,
+        @ICommandService private readonly _commandService: ICommandService
     ) {
         super();
         this._initialize();
@@ -35,6 +40,32 @@ export class FormatPainterRenderController extends Disposable implements IRender
 
     private _initialize() {
         this._bindFormatPainterStatus();
+        this._commandExecutedListener();
+    }
+
+    private _commandExecutedListener() {
+        this.disposeWithMe(this._selectionRenderService.selectionMoveEnd$.subscribe((selections) => {
+            if (this._formatPainterService.getStatus() !== FormatPainterStatus.OFF) {
+                const { rangeWithCoord } = selections[selections.length - 1];
+                this._commandService.executeCommand(ApplyFormatPainterCommand.id, {
+                    unitId: this._context.unitId,
+                    subUnitId: this._context.unit.getActiveSheet()?.getSheetId() || '',
+                    range: {
+                        startRow: rangeWithCoord.startRow,
+                        startColumn: rangeWithCoord.startColumn,
+                        endRow: rangeWithCoord.endRow,
+                        endColumn: rangeWithCoord.endColumn,
+                    },
+                });
+
+                    // if once, turn off the format painter
+                if (this._formatPainterService.getStatus() === FormatPainterStatus.ONCE) {
+                    this._commandService.executeCommand(SetOnceFormatPainterCommand.id);
+                }
+            }
+        })
+
+        );
     }
 
     private _bindFormatPainterStatus() {
